requirements.txt
pulsar-client==2.10.2
confluent-kafka[avro,json,protobuf]>=1.4.2
fastavro
python-dotenv